tickets/src/events/listeners/orderCancelledListener.ts
import {
  Listener,
  Subjects,
  OrderCancelledEvent,
} from "@johnneychentix/common";
import { Error } from "mongoose";
import { Message } from "node-nats-streaming";

import { Ticket } from "../../models/Ticket";
import { TicketEditedPublisher } from "../publishers/ticketEditedPublisher";
import { queueGroupName } from "./queueGroupName";

export class OrderCancelledListener extends Listener<OrderCancelledEvent> {
  readonly subject = Subjects.OrderCancelled;
  queueGroupName = queueGroupName;
  async onMessage(data: OrderCancelledEvent["data"], msg: Message) {
    const ticket = await Ticket.findById(data.ticket.id);

    if (!ticket) {
      throw new Error("Ticket not found");
    }

    ticket.set({ orderId: undefined });
    await ticket.save();

    await new TicketEditedPublisher(this.client).publish({
      id: ticket.id,
      version: ticket.version,
      title: ticket.title,
      price: ticket.price,
      userId: ticket.userId,
    });

    msg.ack();
  }
}
